Renamed Status.jsx to StatusQA.jsx
