+++ ans2ask-ui/src/components/QuestionsOrAnswers/QuestionsOrAnswers.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Options from "../../utils/OptionsQA.jsx"
-import Status from "../../utils/Status.jsx"
+import Status from "../../utils/StatusQA.jsx"
 import "./QuestionsOrAnswers.css";
 
 export default function QuestionsOrAnswers({selectedOption, handleSetSelectedOption}) {
